azure-iot-hub-extractor/pyproject.toml
[tool.poetry]
name = "iot_hub_extractor"
version = "1.0.0"
description = "An Azure IOT Hub extractor"
authors = ["Murad Sæther <murad.sater@cognite.com>"]

[tool.poetry.dependencies]
python = "^3.7"
cognite-extractor-utils = "^1.5.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
csv_extractor = "met_extractor.__main__:main"
